implemented separate numba and numpy internal pf versions
