+++ pandapipes/component_models/abstract_models/branch_models.py
@@ -5,7 +5,7 @@
 import numpy as np
 
 from pandapipes.component_models.abstract_models import Component
-from pandapipes.pf.derivative_toolbox import calc_der_lambda, calc_lambda
+from pandapipes.pf.derivative_calculation import calc_lambda, calc_der_lambda
 from pandapipes.constants import NORMAL_PRESSURE, GRAVITATION_CONSTANT, NORMAL_TEMPERATURE, \
     P_CONVERSION
 from pandapipes.idx_branch import FROM_NODE, TO_NODE, LENGTH, D, TINIT, AREA, K, RHO, ETA, \
@@ -369,152 +369,3 @@ class BranchComponent(Component):
         res_table["mdot_from_kg_per_s"].values[placement_table] = mf_sum / internal_pipes
         res_table["vdot_norm_m3_per_s"].values[placement_table] = vf_sum / internal_pipes
         return placement_table, res_table, branch_pit, node_pit
-
-
-def calculate_derivatives_hydraulic(net, branch_pit, node_pit, options):
-    """
-    Function which creates derivatives.
-
-    :param net: The pandapipes network
-    :type net: pandapipesNet
-    :param branch_pit:
-    :type branch_pit:
-    :param node_pit:
-    :type node_pit:
-    :param idx_lookups:
-    :type idx_lookups:
-    :param options:
-    :type options:
-    :return: No Output.
-    """
-    fluid = get_fluid(net)
-    gas_mode = fluid.is_gas
-    friction_model = options["friction_model"]
-
-    dummy = (branch_pit[:, LENGTH] != 0).astype(np.float64)
-    lambda_, re = calc_lambda(
-        branch_pit[:, VINIT], branch_pit[:, ETA], branch_pit[:, RHO], branch_pit[:, D],
-        branch_pit[:, K], gas_mode, friction_model, dummy, options)
-    der_lambda = calc_der_lambda(branch_pit[:, VINIT], branch_pit[:, ETA], branch_pit[:, RHO],
-                                 branch_pit[:, D], branch_pit[:, K], friction_model, lambda_)
-    branch_pit[:, RE] = re
-    branch_pit[:, LAMBDA] = lambda_
-    from_nodes = branch_pit[:, FROM_NODE].astype(np.int32)
-    to_nodes = branch_pit[:, TO_NODE].astype(np.int32)
-    branch_pit[:, TINIT] = (node_pit[from_nodes, TINIT_NODE] + node_pit[to_nodes, TINIT_NODE]) / 2
-    height_difference = node_pit[from_nodes, HEIGHT] - node_pit[to_nodes, HEIGHT]
-    p_init_i_abs = node_pit[from_nodes, PINIT] + node_pit[from_nodes, PAMB]
-    p_init_i1_abs = node_pit[to_nodes, PINIT] + node_pit[to_nodes, PAMB]
-
-    if not gas_mode:
-        load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1 = derivatives_hydraulic_incomp(
-            branch_pit, der_lambda, p_init_i_abs, p_init_i1_abs, height_difference)
-    else:
-        p_m, der_p_m, der_p_m1 = calc_medium_pressure_with_derivative(p_init_i_abs, p_init_i1_abs)
-        comp_fact = get_fluid(net).get_compressibility(p_m)
-        der_comp = get_fluid(net).get_der_compressibility() * der_p_m
-        der_comp1 = get_fluid(net).get_der_compressibility() * der_p_m1
-        load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1 = derivatives_hydraulic_comp(
-            branch_pit, lambda_, der_lambda, p_init_i_abs, p_init_i1_abs, height_difference,
-            comp_fact, der_comp, der_comp1)
-
-    # return load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1
-
-    branch_pit[:, LOAD_VEC_BRANCHES] = load_vec
-    branch_pit[:, JAC_DERIV_DV] = df_dv
-    branch_pit[:, JAC_DERIV_DP] = df_dp
-    branch_pit[:, JAC_DERIV_DP1] = df_dp1
-    branch_pit[:, LOAD_VEC_NODES] = load_vec_nodes
-    branch_pit[:, JAC_DERIV_DV_NODE] = df_dv_nodes
-
-
-@jit(nopython=True, cache=False)
-def calc_medium_pressure_with_derivative(p_init_i_abs, p_init_i1_abs):
-    p_m = p_init_i_abs.copy()
-    der_p_m = np.ones_like(p_init_i_abs)
-    der_p_m1 = np.ones_like(p_init_i_abs) * (-1)
-    val = 2 / 3
-    for i in range(p_init_i_abs.shape[0]):
-        if p_init_i_abs[i] != p_init_i1_abs[i]:
-            diff_p_sq = p_init_i_abs[i] ** 2 - p_init_i1_abs[i] ** 2
-            diff_p_sq_div = np.divide(1, diff_p_sq)
-            diff_p_cub = p_init_i_abs[i] ** 3 - p_init_i1_abs[i] ** 3
-            p_m[i] = val * diff_p_cub * diff_p_sq_div
-            der_p_m[i] = (3 * p_init_i_abs[i] ** 2 * diff_p_sq - 2 * p_init_i_abs[i] * diff_p_cub) \
-                * diff_p_sq_div ** 2 * val
-            der_p_m1[i] = (-3 * p_init_i1_abs[i] ** 2 * diff_p_sq
-                           + 2 * p_init_i1_abs[i] * diff_p_cub) * diff_p_sq_div ** 2 * val
-    return p_m, der_p_m, der_p_m1
-
-
-@jit(nopython=True, cache=False)
-def derivatives_hydraulic_incomp(branch_pit, der_lambda, p_init_i_abs, p_init_i1_abs,
-                                 height_difference):
-    le = der_lambda.shape[0]
-    load_vec = np.zeros_like(der_lambda)
-    df_dv = np.zeros_like(der_lambda)
-    df_dp = np.ones_like(der_lambda) * (-1)
-    df_dp1 = np.ones_like(der_lambda)
-    load_vec_nodes = np.zeros_like(der_lambda)
-    df_dv_nodes = np.zeros_like(der_lambda)
-
-    for i in range(le):
-        v_init_abs = np.abs(branch_pit[i][VINIT])
-        v_init2 = v_init_abs * branch_pit[i][VINIT]
-        lambda_term = np.divide(branch_pit[i][LENGTH] * branch_pit[i][LAMBDA], branch_pit[i][D]) \
-                      + branch_pit[i][LC]
-        const_p_term = np.divide(branch_pit[i][RHO], P_CONVERSION * 2)
-        df_dv[i] = const_p_term * (2 * v_init_abs * lambda_term + der_lambda[i]
-                                   * np.divide(branch_pit[i][LENGTH], branch_pit[i][D]) * v_init2)
-        load_vec[i] = p_init_i_abs[i] - p_init_i1_abs[i] + branch_pit[i][PL] \
-            + const_p_term * (GRAVITATION_CONSTANT * 2 * height_difference[i]
-                              - v_init2 * lambda_term)
-        mass_flow_dv = branch_pit[i][RHO] * branch_pit[i][AREA]
-        df_dv_nodes[i] = mass_flow_dv
-        load_vec_nodes[i] = mass_flow_dv * branch_pit[i][VINIT]
-    return load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1
-
-
-@jit(nopython=True, cache=False)
-def derivatives_hydraulic_comp(branch_pit, lambda_, der_lambda, p_init_i_abs, p_init_i1_abs,
-                               height_difference, comp_fact, der_comp, der_comp1):
-    le = lambda_.shape[0]
-    load_vec = np.zeros_like(lambda_)
-    df_dv = np.zeros_like(lambda_)
-    df_dp = np.zeros_like(lambda_)
-    df_dp1 = np.zeros_like(lambda_) * (-1)
-    load_vec_nodes = np.zeros_like(der_lambda)
-    df_dv_nodes = np.zeros_like(der_lambda)
-
-    # Formulas for gas pressure loss according to laminar version described in STANET 10
-    # manual, page 1623
-    for i in range(le):
-        # compressibility settings
-        v_init_abs = np.abs(branch_pit[i][VINIT])
-        v_init2 = branch_pit[i][VINIT] * v_init_abs
-        p_diff = p_init_i_abs[i] - p_init_i1_abs[i]
-        p_sum = p_init_i_abs[i] + p_init_i1_abs[i]
-        p_sum_div = np.divide(1, p_sum)
-
-        const_lambda = np.divide(NORMAL_PRESSURE * branch_pit[i][RHO] * branch_pit[i][TINIT],
-                                 NORMAL_TEMPERATURE * P_CONVERSION)
-        const_height = np.divide(
-            branch_pit[i][RHO] * NORMAL_TEMPERATURE * GRAVITATION_CONSTANT * height_difference[i],
-            2 * NORMAL_PRESSURE * branch_pit[i][TINIT] * P_CONVERSION)
-        friction_term = np.divide(lambda_[i] * branch_pit[i][LENGTH],
-                                  branch_pit[i][D]) + branch_pit[i][LC]
-
-        load_vec[i] = p_diff + branch_pit[i][PL] + const_height * p_sum \
-            - const_lambda * comp_fact[i] * v_init2 * friction_term * p_sum_div
-
-        p_deriv = const_lambda * v_init2 * friction_term * p_sum_div
-        df_dp[i] = -1. - p_deriv * (der_comp[i] - comp_fact[i] * p_sum_div) + const_height
-        df_dp1[i] = 1. - p_deriv * (der_comp1[i] - comp_fact[i] * p_sum_div) + const_height
-
-        df_dv[i] = np.divide(2 * const_lambda * comp_fact[i], p_sum) * v_init_abs * friction_term\
-            + np.divide(const_lambda * comp_fact[i] * der_lambda[i] * branch_pit[i][LENGTH]
-                        * v_init2, p_sum * branch_pit[i][D])
-        mass_flow_dv = branch_pit[i][RHO] * branch_pit[i][AREA]
-        df_dv_nodes[i] = mass_flow_dv
-        load_vec_nodes[i] = mass_flow_dv * branch_pit[i][VINIT]
-    return load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1
