+++ pandapipes/pf/derivative_toolbox.py
@@ -2,116 +2,114 @@
 # and Energy System Technology (IEE), Kassel, and University of Kassel. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
-from numpy import linalg
 import numpy as np
-
-try:
-    from numba import jit
-except ImportError:
-    from pandapower.pf.no_numba import jit
-
-
-def calc_lambda(v, eta, rho, d, k, gas_mode, friction_model, dummy, options):
-    """
-    Function calculates the friction factor of a pipe. Turbulence is calculated based on
-    Nikuradse. If v equals 0, a value of 0.001 is used in order to avoid division by zero.
-    This should not be a problem as the pressure loss term will equal zero (lambda * u^2).
-
-    :param v:
-    :type v:
-    :param eta:
-    :type eta:
-    :param rho:
-    :type rho:
-    :param d:
-    :type d:
-    :param k:
-    :type k:
-    :param gas_mode:
-    :type gas_mode:
-    :param friction_model:
-    :type friction_model:
-    :param dummy:
-    :type dummy:
-    :param options:
-    :type options:
-    :return:
-    :rtype:
-    """
-    if gas_mode:
-        re, lambda_laminar, lambda_nikuradse = calc_lambda_nikuradse_comp(v, d, k, eta, rho)
-    else:
-        re, lambda_laminar, lambda_nikuradse = calc_lambda_nikuradse_incomp(v, d, k, eta, rho)
-
-    if friction_model == "colebrook":
-        # TODO: move this import to top level if possible
-        from pandapipes.pipeflow import PipeflowNotConverged
-        max_iter = options.get("max_iter_colebrook", 100)
-        converged, lambda_colebrook = colebrook_nb(re, d, k, lambda_nikuradse, dummy, max_iter)
-        if not converged:
-            raise PipeflowNotConverged(
-                "The Colebrook-White algorithm did not converge. There might be model "
-                "inconsistencies. The maximum iterations can be given as 'max_iter_colebrook' "
-                "argument to the pipeflow.")
-        return lambda_colebrook, re
-    elif friction_model == "swamee-jain":
-        lambda_swamee_jain = 0.25 / ((np.log10(k/(3.7*d) + 5.74/(re**0.9)))**2)
-        return lambda_swamee_jain, re
-    else:
-        # lambda_tot = np.where(re > 2300, lambda_laminar + lambda_nikuradse, lambda_laminar)
-        lambda_tot = lambda_laminar + lambda_nikuradse
-        return lambda_tot, re
+from numpy import linalg
+from pandapipes.constants import P_CONVERSION, GRAVITATION_CONSTANT, NORMAL_PRESSURE, \
+    NORMAL_TEMPERATURE
+from pandapipes.idx_branch import LENGTH, LAMBDA, D, LOSS_COEFFICIENT as LC, RHO, PL, AREA, TINIT, \
+    VINIT
 
 
-def calc_der_lambda(v, eta, rho, d, k, friction_model, lambda_pipe):
-    """
-    Function calculates the derivative of lambda with respect to v. Turbulence is calculated based
-    on Nikuradse. This should not be a problem as the pressure loss term will equal zero
-    (lambda * u^2).
-
-    :param v:
-    :type v:
-    :param eta:
-    :type eta:
-    :param rho:
-    :type rho:
-    :param d:
-    :type d:
-    :param k:
-    :type k:
-    :param friction_model:
-    :type friction_model:
-    :param lambda_pipe:
-    :type lambda_pipe:
-    :return:
-    :rtype:
-    """
-
-    v_corr = np.where(v == 0, 0.00001, v)
-
-    if friction_model == "colebrook":
-        b_term = 2.51 * eta / (rho * d * np.sqrt(lambda_pipe) * v_corr) + k / (3.71 * d)
-
-        df_dv = -2 * (2.51 * eta / (rho * np.sqrt(lambda_pipe) * v_corr ** 2)) \
-                / (np.log(10) * b_term)
-
-        df_dlambda = -0.5 * lambda_pipe ** (-3 / 2) - ((2.51 * eta) / (rho * d * v_corr)) \
-                     * lambda_pipe ** (-3 / 2) * 1 / (np.log(10) * b_term)
-
-        lambda_colebrook_der = df_dv / df_dlambda
-
-        return lambda_colebrook_der
-    elif friction_model == "swamee-jain":
-        param = k/(3.7*d) + 5.74 * (np.abs(eta))**0.9 / ((np.abs(rho*v_corr*d))**0.9)
-        lambda_swamee_jain_der = 0.5/np.log(10) * 1/(np.log(param) ** 3) * 1/param * 5.166 \
-                                 * (np.abs(eta))**0.9/(((np.abs((rho*d)))**0.9) * (np.abs(v_corr))**1.9)
-        return lambda_swamee_jain_der
-    else:
-        lambda_laminar_der = -(64 * eta) / (rho * v_corr ** 2 * d)
-        return lambda_laminar_der
+def derivatives_hydraulic_incomp_np(branch_pit, der_lambda, p_init_i_abs, p_init_i1_abs,
+                                    height_difference):
+    v_init_abs = np.abs(branch_pit[:, VINIT])
+    v_init2 = v_init_abs * branch_pit[:, VINIT]
+    lambda_term = np.divide(branch_pit[:, LENGTH] * branch_pit[:, LAMBDA], branch_pit[:, D]) \
+                  + branch_pit[:, LC]
+    const_p_term = np.divide(branch_pit[:, RHO], P_CONVERSION * 2)
+    df_dv = const_p_term * (2 * v_init_abs * lambda_term + der_lambda
+                            * np.divide(branch_pit[:, LENGTH], branch_pit[:, D]) * v_init2)
+    load_vec = p_init_i_abs - p_init_i1_abs + branch_pit[:, PL] \
+               + const_p_term * (GRAVITATION_CONSTANT * 2 * height_difference
+                                 - v_init2 * lambda_term)
+    mass_flow_dv = branch_pit[:, RHO] * branch_pit[:, AREA]
+    df_dv_nodes = mass_flow_dv
+    load_vec_nodes = mass_flow_dv * branch_pit[:, VINIT]
+    df_dp = np.ones_like(der_lambda) * (-1)
+    df_dp1 = np.ones_like(der_lambda)
+    return load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1
 
 
-def colebrook(re, d, k, lambda_nikuradse, dummy, max_iter):
+def derivatives_hydraulic_comp_np(branch_pit, lambda_, der_lambda, p_init_i_abs, p_init_i1_abs,
+                                  height_difference, comp_fact, der_comp, der_comp1):
+    # Formulas for gas pressure loss according to laminar version described in STANET 10
+    # manual, page 1623
+    v_init_abs = np.abs(branch_pit[:, VINIT])
+    v_init2 = branch_pit[:, VINIT] * v_init_abs
+    p_diff = p_init_i_abs - p_init_i1_abs
+    p_sum = p_init_i_abs + p_init_i1_abs
+    p_sum_div = np.divide(1, p_sum)
+
+    const_lambda = np.divide(NORMAL_PRESSURE * branch_pit[:, RHO] * branch_pit[:, TINIT],
+                             NORMAL_TEMPERATURE * P_CONVERSION)
+    const_height = np.divide(
+        branch_pit[:, RHO] * NORMAL_TEMPERATURE * GRAVITATION_CONSTANT * height_difference,
+        2 * NORMAL_PRESSURE * branch_pit[:, TINIT] * P_CONVERSION)
+    friction_term = np.divide(lambda_ * branch_pit[:, LENGTH], branch_pit[:, D]) + branch_pit[:, LC]
+
+    load_vec = p_diff + branch_pit[:, PL] + const_height * p_sum \
+               - const_lambda * comp_fact * v_init2 * friction_term * p_sum_div
+
+    p_deriv = const_lambda * v_init2 * friction_term * p_sum_div
+    df_dp = -1. - p_deriv * (der_comp - comp_fact * p_sum_div) + const_height
+    df_dp1 = 1. - p_deriv * (der_comp1 - comp_fact * p_sum_div) + const_height
+
+    df_dv = np.divide(2 * const_lambda * comp_fact, p_sum) * v_init_abs * friction_term \
+            + np.divide(const_lambda * comp_fact * der_lambda * branch_pit[:, LENGTH] * v_init2,
+                        p_sum * branch_pit[:, D])
+    mass_flow_dv = branch_pit[:, RHO] * branch_pit[:, AREA]
+    df_dv_nodes = mass_flow_dv
+    load_vec_nodes = mass_flow_dv * branch_pit[:, VINIT]
+
+    return load_vec, load_vec_nodes, df_dv, df_dv_nodes, df_dp, df_dp1
+
+
+def calc_lambda_nikuradse_incomp_np(v, d, k, eta, rho):
+    v_abs = np.abs(v)
+    v_abs[v_abs < 1e-6] = 1e-6
+    re = np.divide(rho * v_abs * d, eta)
+    lambda_laminar = np.zeros_like(v)
+    lambda_laminar[v != 0] = 64 / re[v != 0]
+    lambda_nikuradse = np.divide(1, (-2 * np.log10(k / (3.71 * d))) ** 2)
+    return re, lambda_laminar, lambda_nikuradse
+
+
+def calc_lambda_nikuradse_comp_np(v, d, k, eta, rho):
+    v_abs = np.abs(v)
+    v_abs[v_abs < 1e-6] = 1e-6
+    re = np.divide(rho * v_abs * d, eta)
+    lambda_laminar = np.zeros_like(v)
+    lambda_laminar[v != 0] = 64 / re[v != 0]
+    lambda_nikuradse = np.divide(1, (2 * np.log10(d / k) + 1.14) ** 2)
+    return re, lambda_laminar, lambda_nikuradse
+
+
+def calc_medium_pressure_with_derivative_np(p_init_i_abs, p_init_i1_abs):
+    val = 2 / 3
+    p_m = p_init_i_abs.copy()
+    der_p_m = np.ones_like(p_init_i_abs)
+    der_p_m1 = np.ones_like(p_init_i_abs) * (-1)
+    p_differs = p_init_i_abs != p_init_i1_abs
+
+    if not np.any(p_differs):
+        return p_m, der_p_m, der_p_m1
+
+    p_sq = p_init_i_abs[p_differs] ** 2
+    p1_sq = p_init_i1_abs[p_differs] ** 2
+    diff_p_sq = p_sq - p1_sq
+    diff_p_sq_div = np.divide(1, diff_p_sq)
+    diff_p_cub = p_init_i_abs[p_differs] ** 3 - p_init_i1_abs[p_differs] ** 3
+    factor = diff_p_sq_div ** 2 * val
+
+    p_m[p_differs] = val * diff_p_cub * diff_p_sq_div
+    der_p_m[p_differs] = (3 * p_sq * diff_p_sq - 2 * p_init_i_abs[p_differs] * diff_p_cub) * factor
+    der_p_m1[p_differs] = (-3 * p1_sq * diff_p_sq + 2 * p_init_i1_abs[p_differs] * diff_p_cub) \
+                          * factor
+
+    return p_m, der_p_m, der_p_m1
+
+
+def colebrook_np(re, d, k, lambda_nikuradse, dummy, max_iter):
     """
 
     :param re:
@@ -150,76 +148,6 @@ def colebrook(re, d, k, lambda_nikuradse, dummy, max_iter):
         error_lambda.append(linalg.norm(dx) / (len(dx)))
 
         if error_lambda[niter] <= 1e-4:
-            converged = True
-
-        niter += 1
-
-    return converged, lambda_cb
-
-
-@jit(nopython=True)
-def calc_lambda_nikuradse_comp(v, d, k, eta, rho):
-    lambda_nikuradse = np.empty_like(v)
-    lambda_laminar = np.zeros_like(v)
-    re = np.empty_like(v)
-    v_abs = np.abs(v)
-    for i in range(v.shape[0]):
-        if v_abs[i] < 1e-6:
-            re[i] = np.divide(rho[i] * 1e-6 * d[i], eta[i])
-        else:
-            re[i] = np.divide(rho[i] * v_abs[i] * d[i], eta[i])
-        if v[i] != 0:
-            lambda_laminar[i] = 64 / re[i]
-        lambda_nikuradse[i] = np.divide(1, (2 * np.log10(d[i] / k[i]) + 1.14) ** 2)
-    return re, lambda_laminar, lambda_nikuradse
-
-
-@jit(nopython=True)
-def calc_lambda_nikuradse_incomp(v, d, k, eta, rho):
-    lambda_nikuradse = np.empty_like(v)
-    lambda_laminar = np.zeros_like(v)
-    re = np.empty_like(v)
-    v_abs = np.abs(v)
-    for i in range(v.shape[0]):
-        if v_abs[i] < 1e-6:
-            re[i] = np.divide(rho[i] * 1e-6 * d[i], eta[i])
-        else:
-            re[i] = np.divide(rho[i] * v_abs[i] * d[i], eta[i])
-        if v[i] != 0:
-            lambda_laminar[i] = 64 / re[i]
-        lambda_nikuradse[i] = np.divide(1, (-2 * np.log10(k[i] / (3.71 * d[i]))) ** 2)
-    return re, lambda_laminar, lambda_nikuradse
-
-
-@jit(nopython=True)
-def colebrook_nb(re, d, k, lambda_nikuradse, dummy, max_iter):
-    lambda_cb = lambda_nikuradse.copy()
-    lambda_cb_old = np.empty_like(lambda_nikuradse)
-    converged = False
-    niter = 0
-    factor = np.log(10) * 2.51
-
-    # Inner Newton-loop for calculation of lambda
-    while not converged and niter < max_iter:
-        for i in range(lambda_cb.shape[0]):
-            sqt = np.sqrt(lambda_cb[i])
-            add_val = np.divide(k[i], (3.71 * d[i]))
-            sqt_div = np.divide(1, sqt)
-            re_div = np.divide(1, re[i])
-            sqt_div3 = sqt_div ** 3
-
-            f = sqt_div + 2 * np.log10(2.51 * re_div * sqt_div + add_val)
-            df_dlambda_cb = - 0.5 * sqt_div3 - 2.51 * re_div * sqt_div3 * np.divide(
-                re[i] * sqt + add_val, factor)
-            x = - f / df_dlambda_cb
-
-            lambda_cb_old[i] = lambda_cb[i]
-            lambda_cb[i] += x
-
-        dx = np.abs(lambda_cb - lambda_cb_old) * dummy
-        error_lambda = linalg.norm(dx) / dx.shape[0]
-
-        if error_lambda <= 1e-4:
             converged = True
 
         niter += 1
